PMM-7574 move SecretToggler to shared
