+++ public/app/percona/shared/SecretToggler/SecretToggler.styles.ts
