+++ public/app/percona/shared/SecretToggler/SecretToggler.tsx
