+++ public/app/percona/shared/SecretToggler/SecretToggler.types.ts
