+++ public/app/percona/shared/SecretToggler/index.ts
